Fikset: Kunne ikke slette blokker uten beskyttelse

+++ src/main/no/insane/insane/handlers/BlockProtect.java
@@ -59,7 +59,7 @@ public class BlockProtect {
 		int uid = this.userHandler.getUID(p);
 		if(isProtected(b)) {
 			int owner = getOwner(b);
-			if(owner == uid) {
+			if(owner == uid && (uid != -1)) {
 				if(sqlHandler.update("DELETE FROM blocks WHERE x='"+b.getX()+"' AND y='"+b.getY()+"' AND z='"+b.getZ()+"' AND world='"+b.getWorld().getName()+"'")) {
 					return true;
 				} else {
